+++ cg/services/order_validation_service/models/errors.py
@@ -102,10 +102,15 @@ class InvalidGenePanelsError(CaseError):
 
 
 class InvalidBufferError(CaseSampleError):
-    field: str = "buffer"
+    field: str = "elution_buffer"
     message: str = "The chosen buffer is not allowed when skipping reception control"
 
 
 class RepeatedGenePanelsError(CaseError):
     field: str = "panels"
     message: str = "Gene panels must be unique"
+
+
+class SubjectIdSameAsSampleNameError(CaseSampleError):
+    field: str = "subject_id"
+    message: str = "Subject id must be different from the sample name"
