+++ cg/services/order_validation_service/workflows/tomte/validation_rules.py
@@ -10,6 +10,7 @@ from cg.services.order_validation_service.validators.data.rules import (
 from cg.services.order_validation_service.validators.inter_field.rules import (
     validate_application_compatibility,
     validate_buffer_skip_rc_condition,
+    validate_subject_ids_different_from_sample_names,
     validate_ticket_number_required_if_connected,
 )
 from cg.services.order_validation_service.workflows.tomte.validation.inter_field.rules import (
@@ -43,5 +44,6 @@ TOMTE_CASE_SAMPLE_RULES = [
     validate_fathers_in_same_case_as_children,
     validate_mothers_are_female,
     validate_sample_names_not_repeated,
+    validate_subject_ids_different_from_sample_names,
     validate_wells_contain_at_most_one_sample,
 ]
